requirements.txt
python-telegram-bot==20.2
matplotlib
requests
